+++ src/pipecat/services/openai/base_llm.py
@@ -37,8 +37,8 @@ from pipecat.processors.aggregators.openai_llm_context import (
 )
 from pipecat.processors.frame_processor import FrameDirection
 from pipecat.services.llm_service import FunctionCallFromLLM, LLMService
+from pipecat.utils.asyncio.watchdog_async_iterator import WatchdogAsyncIterator
 from pipecat.utils.tracing.service_decorators import traced_llm
-from pipecat.utils.watchdog_async_iterator import WatchdogAsyncIterator
 
 
 class BaseOpenAILLMService(LLMService):
