+++ src/pipecat/services/openai_realtime_beta/openai.py
@@ -51,9 +51,9 @@ from pipecat.processors.frame_processor import FrameDirection
 from pipecat.services.llm_service import FunctionCallFromLLM, LLMService
 from pipecat.services.openai.llm import OpenAIContextAggregatorPair
 from pipecat.transcriptions.language import Language
+from pipecat.utils.asyncio.watchdog_async_iterator import WatchdogAsyncIterator
 from pipecat.utils.time import time_now_iso8601
 from pipecat.utils.tracing.service_decorators import traced_openai_realtime, traced_stt
-from pipecat.utils.watchdog_async_iterator import WatchdogAsyncIterator
 
 from . import events
 from .context import (
