+++ src/pipecat/utils/asyncio/watchdog_async_iterator.py
@@ -7,7 +7,7 @@
 import asyncio
 from typing import AsyncIterator, Optional
 
-from pipecat.utils.watchdog_reseter import WatchdogReseter
+from pipecat.utils.asyncio.watchdog_reseter import WatchdogReseter
 
 
 class WatchdogAsyncIterator:
